BAEL-8484: Merge articles on Java Collectors
